doit: move stats to top of email and add --attachment
This was just as much extra work as I'd imagined.  We now generate a temp
script to run in the background which collects all the interesting stats
and command output, then assembles it the way we want and sends the email
at the very end.  Way more complicated to code than a simple shell pipeline
like we had before, but also way more flexible now that it's written.